change to SoC pattern
