+++ src/app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import NavWrapper from "../components/nav/NavWrapper";
+import NavWrapper from "../../components/nav/NavWrapper";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
     return <>
