+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import NavWrapper from "../components/nav/NavWrapper";
+import NavWrapper from "../../components/nav/NavWrapper";
 
 export default function Dashboard() {
     return <><h1>Dashboard</h1></>
