+++ src/components/avatar/Avatar.cy.tsx
