+++ src/components/tick-svg/TickSVG.cy.tsx
