+++ src/components/tick-svg/TickSVG.tsx
