+++ src/stories/avatar/Avatar.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Avatar from "../../app/components/avatar/Avatar";
+import Avatar from "../../components/avatar/Avatar";
 
 const meta = {
   title: "Example/Avatar",
