+++ src/stories/button/Button.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Button from "../../app/components/button/Button";
+import Button from "../../components/button/Button";
 
 const meta = {
   title: "Example/Button",
