+++ src/stories/footer/Footer.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Footer from "../../app/components/footer/Footer";
+import Footer from "../../components/footer/Footer";
 
 const meta = {
     title: "Example/Footer",
